Тесты отменяются, настроил непонятно что

+++ Vintello.Services/Services/UserService.cs
@@ -26,8 +26,11 @@ public class UserService(IUserRepository repo) : IUserService
         return RetrievedUserDto.CreateDto(user);
     }
 
-    public async Task<bool?> UpdateAsync(int id, UpdatedUserDto user)
+    public async Task<bool?> UpdateAsync(int id, UpdatedUserDto user, string userRole, string userId)
     {
+        if (userRole == "Client" && userId != id.ToString())
+            return null;
+        
         User? existing = await repo.RetrieveByIdAsync(id);
         if (existing is null) return null;
         
@@ -38,8 +41,11 @@ public class UserService(IUserRepository repo) : IUserService
         return true;
     }
 
-    public async Task<bool?> DeleteAsync(int id)
+    public async Task<bool?> DeleteAsync(int id, string userRole, string userId)
     {
+        if (userRole == "Client" && userId != id.ToString())
+            return null;
+        
         User? user = await repo.RetrieveByIdAsync(id);
         if (user is null) return null;
         return await repo.DeleteAsync(user);
